Modify URI path to properly construct path for ASE. Remove two model files. TypeScript was failing during runtime. When TypeScript creates final javascript file, it puts the Resource base class out of order and the derived classes cannot reference it
