+++ AppLensV2/app/NgModels/Site.ts
@@ -3,6 +3,20 @@
 module SupportCenter {
     "use strict";
 
+    export abstract class Resource {
+        subscriptionId: string;
+        resourceGroup: string;
+        resourceName: string;
+        resourceInternalStamp: string;
+
+        constructor(subscriptionId: string, resourceGroup: string, resourceName: string, resourceInternalStamp: string) {
+            this.subscriptionId = subscriptionId;
+            this.resourceGroup = resourceGroup;
+            this.resourceName = resourceName;
+            this.resourceInternalStamp = resourceInternalStamp;
+        }
+    }
+
     export class Site extends Resource {
         constructor(
             public name: string,
@@ -23,5 +37,16 @@ module SupportCenter {
         numberOfTriggeredWebJobs: number;
         numberOfSlots: number;
         isLinux: boolean;
+    }
+
+    export class HostingEnvironment extends Resource {
+        constructor(
+            subscriptionId: string,
+            resourceGroup: string,
+            hostingEnvironmentName: string,
+            internalStampName: string
+        ) {
+            super(subscriptionId, resourceGroup, hostingEnvironmentName, internalStampName);
+        }
     }
 }
